+++ changelog.md
@@ -6,6 +6,7 @@
 This version is not released yet and is under active development.
 ```
 
+- \[scoop\] Add support for Scoop on Windows. Closes {issue}`546`.
 - \[mpm\] Fix imports from `click.extra`. Closes {issue}`783`.
 
 ## {gh}`5.6.2 (2022-09-27) <compare/v5.6.1...v5.6.2>`